// product-photo-app-v1/types.ts
export enum Sender {
  User = 'user',
  AI = 'ai',
}

export interface ImageAttachment {
  mimeType: string;
  data: string; // Base64 string
}

export interface GeneratedImage {
  id: string;
  data: string; // Base64
  prompt: string;
  timestamp: number;
}

export interface Message {
  id: string;
  sender: Sender;
  text?: string;
  attachments?: ImageAttachment[]; // Images sent by user
  generatedImages?: GeneratedImage[]; // Images generated by AI
  timestamp: number;
}

export type AspectRatio = '1:1' | '3:4' | '4:3' | '9:16' | '16:9';

export interface Session {
  id: string;
  title: string;
  messages: Message[];
  gallery: GeneratedImage[];
  activeReferenceImage: ImageAttachment | null;
  lastModified: number;
  aspectRatio: AspectRatio;
  numberOfImages: number;
}

export interface GeminiConfig {
  model: string;
}